# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KinConnect API
  description: |
    KinConnect is a comprehensive medication management platform that helps patients and caregivers 
    manage medications, track adherence, and coordinate care within family groups.
    
    ## Features
    - **Medication Management**: Create, update, and track medications with reminders
    - **Drug Search**: Search RxNorm database for drug information, interactions, and images
    - **Family Groups**: Invite and manage family members with role-based access
    - **Patient Profiles**: Comprehensive patient health information management
    - **Audit Logging**: Track all access and modifications for compliance
    
    ## Authentication
    All API endpoints require Firebase Authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Rate Limiting
    API requests are limited to 100 requests per 15-minute window per IP address.
    
    ## Versioning
    Current API version: v1
    
  version: 1.0.0
  contact:
    name: KinConnect Support
    email: support@kinconnect.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/{project-id}/us-central1/api
    description: Local Development Server
    variables:
      project-id:
        default: kinconnect-dev
        description: Firebase Project ID
  - url: https://us-central1-{project-id}.cloudfunctions.net/api
...
    variables:
...
security:
  - bearerAuth: []
paths:
...
  /invitations/send:
    post:
      tags:
        - Invitations
      summary: Send patient invitation
      description: Send an invitation to a patient to join a family group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationInput'
            examples:
              basic:
                value:
                  email: "patient@example.com"
                  patientName: "Jane Smith"
              with_message:
                value:
                  email: "patient@example.com"
                  patientName: "Jane Smith"
                  message: "I'd like to help manage your medications"
      responses:
        '201':
          description: Invitation sent successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invitation'
              examples:
                success:
                  value:
                    success: true
                    data:
                      invitationId: "inv123"
                      inviterUid: "user123"
                      inviterName: "John Doe"
                      inviterEmail: "john@example.com"
                      patientEmail: "patient@example.com"
                      patientName: "Jane Smith"
                      message: "I'd like to help manage your medications"
                      status: "pending"
                      createdAt: "2024-01-01T00:00:00Z"
                      expiresAt: "2024-01-08T00:00:00Z"
                    message: "Invitation sent successfully"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invitations/{invitationId}:
    get:
      tags:
        - Invitations
      summary: Get invitation details
      description: Get details of a specific invitation (no auth required for accepting invitations)
      parameters:
        - name: invitationId
          in: path
          required: true
          description: Invitation ID
          schema:
            type: string
          example: "inv123"
      security: []
      responses:
        '200':
          description: Invitation details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invitation'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '410':
          description: Invitation expired or already accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                expired:
                  value:
                    success: false
                    error: "Invitation has expired"
                accepted:
                  value:
                    success: false
                    error: "Invitation has already been accepted"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invitations/{invitationId}/accept:
    post:
      tags:
        - Invitations
      summary: Accept invitation
      description: Accept an invitation and join the family group
      security:
        - bearerAuth: []
      parameters:
        - name: invitationId
          in: path
          required: true
          description: Invitation ID
          schema:
            type: string
          example: "inv123"
      responses:
        '200':
          description: Invitation accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "Invitation accepted successfully and family group updated"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '410':
          description: Invitation expired or already accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invitations/sent:
    get:
      tags:
        - Invitations
      summary: Get sent invitations
      description: Get all invitations sent by the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sent invitations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invitation'
              examples:
                success:
                  value:
                    success: true
                    data:
                      - id: "inv123"
                        inviterUid: "user123"
                        inviterName: "John Doe"
                        patientEmail: "patient@example.com"
                        patientName: "Jane Smith"
                        status: "pending"
                        createdAt: "2024-01-01T00:00:00Z"
                        expiresAt: "2024-01-08T00:00:00Z"
                    message: "Sent invitations retrieved successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== FAMILY ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained from Firebase Authentication
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message about the result
        error:
          type: string
          description: Error message if request failed
        data:
          description: Response data (type varies by endpoint)

    # ==================== USER & PATIENT ====================
...
    Invitation:
      type: object
      required:
        - inviterUid
        - inviterName
        - inviterEmail
        - patientEmail
        - patientName
        - status
        - createdAt
        - expiresAt
      properties:
        id:
          type: string
          description: Invitation ID
        invitationId:
          type: string
          description: Invitation ID (alternative field name)
        inviterUid:
          type: string
          description: UID of the user sending the invitation
        inviterName:
          type: string
          description: Name of the inviter
        inviterEmail:
          type: string
          format: email
          description: Email of the inviter
        patientEmail:
          type: string
          format: email
          description: Email of the invited patient
        patientName:
          type: string
          description: Name of the invited patient
        message:
          type: string
          description: Optional message from inviter
        status:
          type: string
          enum: [pending, accepted, rejected, expired]
          description: Status of the invitation
        acceptedBy:
          type: string
          description: UID of user who accepted the invitation
        acceptedAt:
          type: string
          format: date-time
          description: Timestamp when invitation was accepted
        createdAt:
          type: string
          format: date-time
          description: Timestamp when invitation was created
        expiresAt:
          type: string
          format: date-time
          description: Timestamp when invitation expires
    InvitationInput:
      type: object
      required:
        - email
        - patientName
      properties:
        email:
          type: string
          format: email
          description: Email address of the patient to invite
        patientName:
          type: string
          description: Full name of the patient
        message:
          type: string
          description: Optional personal message
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            missing_token:
              value:
                success: false
                error: "Authorization header required"
            invalid_token:
              value:
                success: false
                error: "Invalid or expired token"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            not_found:
              value:
                success: false
                error: "Resource not found"
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            validation_error:
              value:
                success: false
                error: "Invalid request parameters"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            server_error:
              value:
                success: false
                error: "Internal server error"